increase maximum auto dock timeout to one month
The auto dock timeout maximum of one hour is very short. An increased maximum makes it possible to idle next to the trashcan for a few hours and wait for the user to empty the bin.
The limit of one month is chosen without any good reason, besides the current code requiring a limit.
Tests show that the robot goes to sleep after about 8 hours, causing higher timeouts to fail. This limit of 8 hours is not a feature but an issue that can be addressed in the future.